Helpers/download_and_filter_sbb_images/requirements.txt
absl-py==0.2.2
astor==0.6.2
astroid==1.6.4
backcall==0.1.0
bleach==1.5.0
colorama==0.3.9
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
gast==0.2.0
grpcio==1.12.0
h5py==2.7.1
html5lib==0.9999999
imutils==0.4.6
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.6
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.5.0
numpy==1.14.3
opencv-contrib-python==3.4.1.15
pandas==0.23.0
pandocfilters==1.4.2
parso==0.2.1
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
Pygments==2.2.0
pylint==1.9.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
pywinpty==0.5.2
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.8.0
tensorflow==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
Theano==1.0.2
tornado==5.0.2
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.10.11
